Separate ErrorMessage out from ErrorBoundary
ErrorBoundary had withStyles, but only used it in the case where it
actually has an error to render. Extract the error message rendering
into a different component, so that an ErrorBoundary doesn't bring along
additional HoCs.
