+++ packages/lesswrong/components/common/ErrorMessage.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import { registerComponent } from 'meteor/vulcan:core';
+import { withStyles } from '@material-ui/core/styles';
+import Typography from '@material-ui/core/Typography';
+
+const styles = theme => ({
+  errorText: {
+    color: theme.palette.error.main,
+  }
+})
+
+const ErrorMessage = ({message, classes}) => {
+  return <Typography
+    className={classes.errorText}
+    align="center"
+    variant="body1"
+  >
+    Error: {message}
+  </Typography>
+}
+
+registerComponent("ErrorMessage", ErrorMessage, withStyles(styles, { name: "ErrorMessage" }));
